separate SwitchedLikelihood tests to own file
