+++ tests/gpflow/likelihoods/test_likelihoods.py
@@ -12,14 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import os
 import numpy as np
 import pytest
 import tensorflow as tf
 from numpy.testing import assert_allclose
 
 import gpflow
-from gpflow.inducing_variables import InducingPoints
 from gpflow.likelihoods import (
     Bernoulli,
     Beta,
@@ -113,7 +111,7 @@ def test_no_missing_likelihoods():
         if likelihood_class in tested_likelihood_types:
             continue  # already tested
         if likelihood_class is SwitchedLikelihood:
-            continue  # tested separately
+            continue  # tested separately, see test_switched_likelihood.py
         if likelihood_class is MonteCarloLikelihood:
             continue  # abstract base class
         if issubclass(likelihood_class, MonteCarloLikelihood):
@@ -380,107 +378,3 @@ def test_robust_max_multiclass_eps_k1_changes(num_classes, initial_epsilon, new_
     expected_eps_k2 = new_epsilon / (num_classes - 1.0)
     actual_eps_k2 = likelihood.eps_k1
     assert_allclose(expected_eps_k2, actual_eps_k2)
-
-
-@pytest.mark.parametrize("Y_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("F_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("Fvar_list", [[tf.exp(tf.random.normal((i, 2))) for i in range(3, 6)]])
-@pytest.mark.parametrize("Y_label", [[tf.ones((i, 2)) * (i - 3.0) for i in range(3, 6)]])
-def test_switched_likelihood_log_prob(Y_list, F_list, Fvar_list, Y_label):
-    """
-    SwitchedLikelihood is separately tested here.
-    Here, we make sure the partition-stitch works fine.
-    """
-    Y_perm = list(range(3 + 4 + 5))
-    np.random.shuffle(Y_perm)
-    # shuffle the original data
-    Y_sw = np.hstack([np.concatenate(Y_list), np.concatenate(Y_label)])[Y_perm, :3]
-    F_sw = np.concatenate(F_list)[Y_perm, :]
-    likelihoods = [Gaussian()] * 3
-    for lik in likelihoods:
-        lik.variance = np.exp(np.random.randn(1)).squeeze().astype(np.float32)
-    switched_likelihood = SwitchedLikelihood(likelihoods)
-
-    switched_results = switched_likelihood.log_prob(F_sw, Y_sw)
-    results = [lik.log_prob(f, y) for lik, y, f in zip(likelihoods, Y_list, F_list)]
-
-    assert_allclose(switched_results, np.concatenate(results)[Y_perm, :])
-
-
-@pytest.mark.parametrize("Y_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("F_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("Fvar_list", [[tf.exp(tf.random.normal((i, 2))) for i in range(3, 6)]])
-@pytest.mark.parametrize("Y_label", [[tf.ones((i, 2)) * (i - 3.0) for i in range(3, 6)]])
-def test_switched_likelihood_predict_density(Y_list, F_list, Fvar_list, Y_label):
-    Y_perm = list(range(3 + 4 + 5))
-    np.random.shuffle(Y_perm)
-    # shuffle the original data
-    Y_sw = np.hstack([np.concatenate(Y_list), np.concatenate(Y_label)])[Y_perm, :3]
-    F_sw = np.concatenate(F_list)[Y_perm, :]
-    Fvar_sw = np.concatenate(Fvar_list)[Y_perm, :]
-
-    likelihoods = [Gaussian()] * 3
-    for lik in likelihoods:
-        lik.variance = np.exp(np.random.randn(1)).squeeze().astype(np.float32)
-    switched_likelihood = SwitchedLikelihood(likelihoods)
-
-    switched_results = switched_likelihood.predict_density(F_sw, Fvar_sw, Y_sw)
-    # likelihood
-    results = [
-        lik.predict_density(f, fvar, y)
-        for lik, y, f, fvar in zip(likelihoods, Y_list, F_list, Fvar_list)
-    ]
-    assert_allclose(switched_results, np.concatenate(results)[Y_perm, :])
-
-
-@pytest.mark.parametrize("Y_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("F_list", [[tf.random.normal((i, 2)) for i in range(3, 6)]])
-@pytest.mark.parametrize("Fvar_list", [[tf.exp(tf.random.normal((i, 2))) for i in range(3, 6)]])
-@pytest.mark.parametrize("Y_label", [[tf.ones((i, 2)) * (i - 3.0) for i in range(3, 6)]])
-def test__switched_likelihood_variational_expectations(Y_list, F_list, Fvar_list, Y_label):
-    Y_perm = list(range(3 + 4 + 5))
-    np.random.shuffle(Y_perm)
-    # shuffle the original data
-    Y_sw = np.hstack([np.concatenate(Y_list), np.concatenate(Y_label)])[Y_perm, :3]
-    F_sw = np.concatenate(F_list)[Y_perm, :]
-    Fvar_sw = np.concatenate(Fvar_list)[Y_perm, :]
-
-    likelihoods = [Gaussian()] * 3
-    for lik in likelihoods:
-        lik.variance = np.exp(np.random.randn(1)).squeeze().astype(np.float32)
-    switched_likelihood = SwitchedLikelihood(likelihoods)
-
-    switched_results = switched_likelihood.variational_expectations(F_sw, Fvar_sw, Y_sw)
-    results = [
-        lik.variational_expectations(f, fvar, y)
-        for lik, y, f, fvar in zip(likelihoods, Y_list, F_list, Fvar_list)
-    ]
-    assert_allclose(switched_results, np.concatenate(results)[Y_perm, :])
-
-
-@pytest.mark.parametrize("num_latent_gps", [1, 2])
-def test_switched_likelihood_regression_valid_num_latent_gps(num_latent_gps):
-    """
-    A Regression test when using Switched likelihood: the number of latent
-    functions in a GP model must be equal to the number of columns in Y minus
-    one. The final column of Y is used to index the switch. If the number of
-    latent functions does not match, an exception will be raised.
-    """
-    x = np.random.randn(100, 1)
-    y = np.hstack((np.random.randn(100, 1), np.random.randint(0, 3, (100, 1))))
-    data = x, y
-
-    Z = InducingPoints(np.random.randn(num_latent_gps, 1))
-    likelihoods = [StudentT()] * 3
-    switched_likelihood = SwitchedLikelihood(likelihoods)
-    m = gpflow.models.SVGP(
-        kernel=gpflow.kernels.Matern12(),
-        inducing_variable=Z,
-        likelihood=switched_likelihood,
-        num_latent_gps=num_latent_gps,
-    )
-    if num_latent_gps == 1:
-        m.log_likelihood(data)
-    else:
-        with pytest.raises(tf.errors.InvalidArgumentError):
-            m.log_likelihood(data)
